ashwin fix for list below
Upload/Manage My Artifacts->The Upload Date field is not populating when a document is uploaded
Character Count->	The character count validation should not be invoked for Save draft. It should be invoked only for Submit task.
Character Count->Character count should be configurable per task. Task1
Upload/Manage My Artifacts->The Upload Date field is not populating when a document is uploaded